DE2/airflow/requirements.txt
# Airflow extras (light: just the GCP provider we need for BigQuery external table)
apache-airflow-providers-google==10.19.0

# Ingestion deps
requests>=2.31
pyarrow>=15.0
google-cloud-storage>=2.14
google-cloud-bigquery>=3.17

# dbt
dbt-bigquery==1.8.2
